requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
faiss-cpu
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
numpy
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
pymupdf
python-multipart==0.0.22
requests
sentence-transformers
SQLAlchemy==2.0.48
starlette==0.52.1
tqdm
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0